Fix mypy errors surfaced by the restored type gate
- maxtp: raise a clear error when no node is eligible for a task, instead of
  passing None into add_task() and crashing confusingly
- inspirit (throughput + online policy): cast numpy floating results to float
- multi_obj / parametric.online: drop redundant re-annotations of existing names
- parametric.online: move the type: ignore[override] onto the def line so it
  actually applies
- estimate_stochastic_scheduler: type forwarded kwargs as Dict[str, Any]
- parametric: documented type: ignore[call-arg] for the ABC+BaseModel MRO
  limitation (pydantic accepts the declared field kwargs at runtime)

All gates green: 96 tests, mypy clean, ruff check/format clean.

+++ src/saga/schedulers/throughput/inspirit.py
@@ -19,7 +19,7 @@ def compute_inspiring_efficiency(
     Returns:
        efficiency_ranks: A dict of task_name to inspiring efficiency, where a task's inspiring efficiency is the number of tasks expected to finish in the given time window.
     """
-    average_network_speed = np.mean([node.speed for node in network.nodes])
+    average_network_speed = float(np.mean([node.speed for node in network.nodes]))
 
     # the paper suggests using the average execution time of tasks of a certain type, but
     # because we don't have task types, we will just manually calculate execution costs of
@@ -35,7 +35,7 @@ def compute_inspiring_efficiency(
         """
         count = 0
         # counter = 0
-        time = 0
+        time = 0.0
         current_gen = {task.name}
         next_gen = set()
         visited = set()
